MOdification club + simili sgbd

+++ club.py
@@ -18,9 +18,22 @@ def lecturerecord():
 
     return [nom, prenom, adresse, cp, tel]
 
+def modificationrecord(record_fichier):
+    while True:
+        print(record_fichier)
+        date = input('Entrer la date de naissance: ')
+        sexe = input('Entrere le sexe: ')
+        print(date + ' ' + sexe)
+        choix = input("Ces informations sont-elles correctes? (n/Y)")
+        if choix == '':
+            break
+    record_fichier = record_fichier + '#' + date + '#' + sexe + '#'
 
-def enregistrementrecord(record):
-    of = open(fname, 'a')
+    return record_fichier
+
+
+def enregistrementrecord(record, nom):
+    of = open(nom, 'a')
     i = 0
     while i < len(record):
         of.write(record[i]+'#')
@@ -29,15 +42,30 @@ def enregistrementrecord(record):
     of.close()
 
 
-#fname = input('Nom du fichier: ')
 fname = 'db.dat'
-if os.path.exists(fname):
-    os.remove(fname)
+fnamebis = 'db2.dat'
 
-while True:
-    newrecord = lecturerecord()
-    if newrecord != "":
-        print("enregitrement")
-        enregistrementrecord(newrecord)
-    else:
-        break
+choix = input('Voulez-vous modifier ou creer un nouveau fichier(1 ou 2):')
+if choix == '1':
+    # fname = input("Nom du fichier d'origine: ")
+    # fnamebis = input("Nom du fichier de copie: ")
+    of = open(fname, 'r')
+    while True:
+        newrecord = modificationrecord(of.readline())
+        if newrecord != "":
+            # if os.path.exists(fnamebis):
+            #    os.remove(fnamebis)
+            enregistrementrecord(newrecord, fnamebis)
+        else:
+            break
+elif choix == '2':
+    # fname = input("Nom du fichier d'origine: ")
+    while True:
+        newrecord = lecturerecord()
+        if newrecord != "":
+            print("enregitrement")
+            if os.path.exists(fname):
+                os.remove(fname)
+            enregistrementrecord(newrecord, fname)
+        else:
+            break
